Decouple inlay hints from Label
Label is used in far more places than previously thought (e.g.
PlugInstall or for default constructors) and the previous accent was too
much.
By default vim links Label to Statement (which is red), so let's define
Label as a variant of red and use the accent color just for the inline
hints.