diffAdded -> Added
With the upstream changes in
https://github.com/vim/vim/commit/124371c5a149a8c0c75c04b6c90ac11e71a0aa97
the diffAdded group already links to Added now (which transitively maps
to DiffAdd from our colorscheme).

Instead overwrite the Added highlight group.

In the future we should reconsider if we really want diffAdded (used for
the diff filetype (e.g. opening a file.diff)) and DiffAdd (used for
actual diffing in vim) to have the same color. The latter one is
expected to have a very strong highlight (which we currently have for
both), but the former might be expected to have a slightly less
invasive/aggressive highlight, e.g. compare it with the commit linked
above.

If we ever modify this, we should make sure that the diffAdded groups
keep working (which it does right now due to transitive mapping over
Added). In that case it is probably easiest to keep vim's existing
default link from diffAdded to Added, then define Added ourself (instead
of linking to DiffAdd as we do now), and still keep defining DiffAdd
ourself.